# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Capital-Challenge backend API.",
        "title": "Capital-Challenge API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
...
        "/user-assets/profits/{user_id}": {
            "get": {
                "description": "get user assets profits",
                "tags": [
                    "user_assets"
                ],
                "summary": "GetUserAssetsProfits",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user_id",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1userassets.UserAssetsProfits"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/user-assets/value/{user_id}": {
            "get": {
                "description": "get user assets values",
                "tags": [
                    "user_assets"
                ],
                "summary": "GetUserAssetsValue",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user_id",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1userassets.UserAssetsValues"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/user-assets/{user_id}": {
            "get": {
                "description": "get user assets for current game",
                "tags": [
                    "user_assets"
                ],
                "summary": "GetUserAssets",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user_id",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.UserAssets"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/users/": {
            "post": {
                "description": "Login user",
                "tags": [
                    "users"
                ],
                "summary": "Login user",
                "parameters": [
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1Users.UserLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/users/register": {
            "post": {
                "description": "Register user",
                "tags": [
                    "users"
                ],
                "summary": "Register user",
                "parameters": [
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1Users.UserRegistrationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1Users.UserRegistrationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.CompanyStock": {
            "type": "object",
            "properties": {
                "close_price": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "highest_price": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "lowest_price": {
                    "type": "number"
                },
                "number_of_transactions": {
                    "type": "integer"
                },
                "open_price": {
                    "type": "number"
                },
                "otc": {
                    "type": "boolean"
                },
                "ticker": {
                    "type": "string"
                },
                "trading_volume": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "volume_weighted_average_price": {
                    "type": "number"
                }
            }
        },
        "models.UserAssets": {
            "type": "object",
            "properties": {
                "company_stock_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "price_per_unit": {
                    "type": "number"
                },
                "quantity": {
                    "type": "integer"
                },
                "ticker": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "v1Users.UserLoginRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "v1Users.UserRegistrationRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "v1Users.UserRegistrationResponse": {
            "type": "object",
            "properties": {
                "current_balance": {
                    "type": "number"
                },
                "email": {
                    "type": "string"
                },
                "starting_balance": {
                    "type": "number"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "v1userassets.UserAssetsProfits": {
            "type": "object",
            "properties": {
                "company_stock": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1userassets.UserAssetsProfitsPerStock"
                    }
                },
                "total_current_value": {
                    "type": "number"
                },
                "total_profit_": {
                    "type": "number"
                },
                "total_profit_margin": {
                    "type": "string"
                },
                "total_spent": {
                    "type": "number"
                }
            }
        },
        "v1userassets.UserAssetsProfitsPerStock": {
            "type": "object",
            "properties": {
                "buy_price": {
                    "type": "number"
                },
                "company_stock": {
                    "$ref": "#/definitions/models.CompanyStock"
                },
                "current_price": {
                    "type": "number"
                },
                "profit_margin": {
                    "type": "string"
                },
                "profit_per_unit": {
                    "type": "number"
                },
                "quantity": {
                    "type": "integer"
                },
                "total_buy_price": {
                    "type": "number"
                },
                "total_current_price": {
                    "type": "number"
                },
                "total_profit": {
                    "type": "number"
                }
            }
        },
        "v1userassets.UserAssetsValues": {
            "type": "object",
            "properties": {
                "quantity": {
                    "type": "integer"
                },
                "ticker": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                },
                "value_per_unit": {
                    "type": "number"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}